refactor(auth): Перенес zod-схему для валидации формы регистрации в отдельный файл и сделал ее экспортируемой.
