+++ features/auth/ui/Register/RegisterForm/RegisterForm.tsx
@@ -7,9 +7,9 @@ import googleIcon from "@/shared/assets/img/google-icon.png"
 import gitHubIcon from "@/shared/assets/img/github-icon.png"
 import { TextField } from "@/shared/ui/text-field/text-field"
 import { Button } from "@/shared/ui"
-import z from "zod"
 import { useForm } from "react-hook-form"
 import { zodResolver } from "@hookform/resolvers/zod"
+import { SignUpInputs, signUpSchema } from "@/shared/lib/signUpSchema/signUpSchema"
 
 export const RegisterForm = () => {
   const {
@@ -82,31 +82,3 @@ export const RegisterForm = () => {
     </div>
   )
 }
-
-const signUpSchema = z
-  .object({
-    userName: z
-      .string()
-      .min(1, "Username is required")
-      .regex(/^[-_A-Za-z0-9]+$/, "Username must contain only numbers and latin letters")
-      .min(6, "Minimum number of characters 6")
-      .max(30, "Maximum number of characters 30"),
-    email: z.email("The email is must match the format example@example.com"),
-    password: z
-      .string()
-      .min(1, "Password is required")
-      .regex(
-        /^(?=.*[A-Z])(?=.*[0-9])(?=.*[a-z])[-_A-Za-z0-9!"#$%&'()*+,./:;<=>?@[\]\\^{|}~]+$/,
-        "Password must contain 0-9, a-z, A-Z, ! \" # $ % & ' ( ) * + , - . / : ; < = > ? @ [ \ ] ^ _ { | } ~"
-      )
-      .min(6, "Minimum number of characters 6")
-      .max(20, "Maximum number of characters 20"),
-    confirmPassword: z.string(),
-    agree: z.boolean().refine((data) => data === true),
-  })
-  .refine((data) => data.password === data.confirmPassword, {
-    message: "Passwords must match",
-    path: ["confirmPassword"],
-  })
-
-export type SignUpInputs = z.infer<typeof signUpSchema>
